ui: Move stories to schedule page
