+++ lib/presentation/core/routes/routes.dart
@@ -7,7 +7,7 @@ import 'package:rtu_mirea_app/presentation/pages/login/login_page.dart';
 import 'package:rtu_mirea_app/presentation/pages/map/map_page.dart';
 import 'package:rtu_mirea_app/presentation/pages/news/news_details_page.dart';
 import 'package:rtu_mirea_app/presentation/pages/news/news_page.dart';
-import 'package:rtu_mirea_app/presentation/pages/news/widgets/stories_wrapper.dart';
+import 'package:rtu_mirea_app/presentation/pages/schedule/widgets/stories_wrapper.dart';
 import 'package:rtu_mirea_app/presentation/pages/onboarding/onboarding_page.dart';
 import 'package:rtu_mirea_app/presentation/pages/profile/about_app_page.dart';
 import 'package:rtu_mirea_app/presentation/pages/profile/profile_announces_page.dart';
