+++ lib/presentation/pages/news/news_page.dart
@@ -1,16 +1,13 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:rtu_mirea_app/domain/entities/news_item.dart';
-import 'package:rtu_mirea_app/domain/entities/story.dart';
 import 'package:rtu_mirea_app/presentation/bloc/news_bloc/news_bloc.dart';
 import 'package:rtu_mirea_app/presentation/bloc/stories_bloc/stories_bloc.dart';
 import 'package:rtu_mirea_app/presentation/widgets/buttons/app_settings_button.dart';
 import 'package:rtu_mirea_app/presentation/widgets/buttons/primary_tab_button.dart';
 import 'package:shimmer/shimmer.dart';
 import 'widgets/news_item.dart';
-import 'widgets/story_item.dart';
 import 'widgets/tag_badge.dart';
 import 'package:rtu_mirea_app/presentation/typography.dart';
 import 'package:rtu_mirea_app/presentation/theme.dart';
@@ -154,38 +151,6 @@ class _NewsPageState extends State<NewsPage> {
     );
   }
 
-  List<Story> _getActualStories(List<Story> stories) {
-    List<Story> actualStories = [];
-    for (final story in stories) {
-      if (DateTime.now().compareTo(story.stopShowDate) == -1) {
-        actualStories.add(story);
-      }
-    }
-
-    return actualStories;
-  }
-
-  Widget _buildFlexibleSpace(List<Story> stories) {
-    return SizedBox(
-      height: 120,
-      child: ListView.separated(
-        padding: const EdgeInsets.symmetric(horizontal: 20),
-        scrollDirection: Axis.horizontal,
-        itemBuilder: (_, int i) {
-          if (DateTime.now().compareTo(stories[i].stopShowDate) == -1) {
-            return StoryWidget(
-              stories: stories,
-              storyIndex: i,
-            );
-          }
-          return Container();
-        },
-        separatorBuilder: (_, int i) => const SizedBox(width: 10),
-        itemCount: stories.length,
-      ),
-    );
-  }
-
   int _getColumnCount(double screenWidth) {
     if (screenWidth < 900) {
       return 1;
@@ -218,28 +183,12 @@ class _NewsPageState extends State<NewsPage> {
         controller: _scrollController,
         headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) {
           return [
-            BlocBuilder<StoriesBloc, StoriesState>(
-              builder: (context, state) {
-                if (state is StoriesInitial) {
-                  context.read<StoriesBloc>().add(LoadStories());
-                } else if (state is StoriesLoaded) {
-                  final actualStories = _getActualStories(state.stories);
-                  if (actualStories.isNotEmpty) {
-                    return SliverAppBar(
-                        automaticallyImplyLeading: false,
-                        expandedHeight: 110,
-                        actionsIconTheme: const IconThemeData(opacity: 0.0),
-                        flexibleSpace: _buildFlexibleSpace(actualStories));
-                  }
-                }
-                return const SliverAppBar(
-                  automaticallyImplyLeading: false,
-                  toolbarHeight: 0,
-                  elevation: 0,
-                  primary: false,
-                  actionsIconTheme: IconThemeData(opacity: 0.0),
-                );
-              },
+            const SliverAppBar(
+              automaticallyImplyLeading: false,
+              toolbarHeight: 0,
+              elevation: 0,
+              primary: false,
+              actionsIconTheme: IconThemeData(opacity: 0.0),
             ),
           ];
         },
